backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docx2txt==0.9
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
filelock==3.20.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.2.0
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-mistralai==1.1.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
numpy==2.4.0
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.1.0
postgrest==2.27.1
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyroaring==1.0.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
realtime==2.27.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
starlette==0.50.0
storage3==2.27.1
streamlit==1.52.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
